Rename folders that export components. Use capitalized names. Refactor 'sample-nested-page' to 'SampleNestedPage' and 'welcome' to 'WelcomePage'
