+++ react-app/src/js/WelcomePage/actions.js
