+++ react-app/src/js/WelcomePage/index.js
