+++ react-app/src/js/WelcomePage/index.test.js
