+++ react-app/src/js/WelcomePage/reducers.js
